caching: Extract out common test cases into base classes.
